マージ

+++ application/Sourcecode/Enemy/IEnemy.cpp
@@ -1,6 +1,12 @@
 #include "IEnemy.h"
 #include "Player.h"
 
+/**
+ * @file IEnemy.cpp
+ * @brief インターフェースクラス
+ * @brief 敵を作る際に継承するクラス
+ */
+
 Player* IEnemy::splayer_ = nullptr;
 
 IEnemy::IEnemy(EnemyType type, bool isGravityImpact, int32_t health)
@@ -69,7 +75,7 @@ void IEnemy::ColPosUpdate()
 	col_.radius = obj_->GetTransform()->scale_.x;
 }
 
-void IEnemy::Damage(const Vector3& knockVec, int32_t damageValue, int32_t cooltime)
+void IEnemy::Damage(const Vector3& knockVec, int32_t damageValue, float cooltime)
 {
 	if (damageCoolTime_.GetIsEnd())
 	{
