+++ application/Sourcecode/Interface/IBullet.h
@@ -7,6 +7,12 @@
 #include "IActor.h"
 #include "Spline.h"
 
+/**
+ * @file IBullet.h
+ * @brief インターフェースクラス
+ * @brief 弾・遠距離魔法を作る際に継承するクラス
+ */
+
 class IBullet
 {
 private:
@@ -15,7 +21,7 @@ private:
 		Sphere col_;
 		int32_t damage = 10;
 		//敵の攻撃当たり判定有効までの時間
-		int32_t damageCoolTime = 20;
+		float damageCoolTime = 20;
 		//ノックバックの強さ
 		Vector3 knockPower = { 0.5f,0.5f,0.5f };
 		//上方向のノックバック
@@ -38,7 +44,7 @@ protected:
 	int32_t costMP_;
 public:
 	IBullet(IActor* selfActor,
-		int32_t maxTime = 20, int32_t damage = 10, int32_t damageCoolTime = 25, int32_t costMP = 10);
+		int32_t maxTime = 20, int32_t damage = 10, float damageCoolTime = 25.f, int32_t costMP = 10);
 	virtual ~IBullet() {};
 	void Update();
 	void DrawCol();
